refactor: split resolvers
